requirements.txt
django
djangorestframework
pylint
pyyaml
faker